Fully decoupled pushy_node_state FSM start from command switch
A deadlock could occur when the command switch started a node state
FSM process in response to a pushy message. The deadlock happened
because the command switch would block waiting on the supervisor
to start the FSM process but the FSM process was sending command
messages via the command switch.

This commit adds a "send after" capability to the command switch
such that we can decouple the FSM from the switch at startup.