Teach pushy how to run under gdb for easier NIF debugging
bin/opscode-pushy-server has a new option, 'dbg'. When the server
is started via this option it runs in console mode with gdb attached.
gdb initialization and startup is automated by a small script in
rel/files/debug.gdb. The intention behind this option is to make
it easier to debug VM crashes due to errors in NIFs.