Making debug mode consistent across pushy and erchef
I moved the debug.gdb file from bin/ to etc/ based on feedback from
Seth F. I agree this is a better home for the file since it isn't
executable on its own. This commit moves debug.gdb into etc for the
pushy server to make it consistent with erchef.