Fix spec for pushy_principal:fetch_principal/2
The spec wasn't accounting for a possible {conn_failed, term()} tuple.

Additionally, it parse_json_response/1 appears that it was intended to
actually throw an error tuple rather than just returning it, based on
its spec and the handling present in fetch_principal/2 and
request_principal/2.

More work could be done here to clean this module up a bit, but the
focus here is getting the code dialyzer clean.