Use $(REBAR) instead of rebar in concrete.mk
In this project, the $(REBAR) variable gets set in the lock_deps.mk
file, which is included in the main Makefile first.  Concrete doesn't
yet natively support the lock_deps stuff, so I'm hacking this locally
so our locked builds behave like they used to; we can work on formally
integrating into concrete somehow in the future.