Filthy hack to get concrete dialyzer working properly
Before, running 'make' in a fresh checkout would not run dialyzer with
all your application's dependencies, because they would not exist in
the deps directory (indeed, the directory would not even exist!) when
the Makefile was evaluated.  Subsequent runs would build a deps.plt
file and use that.

We now use an external script to compute the desired contents of the
deps.plt at target execution time, rather than the initial target
"compilation" time.  I have no doubt that it could be prettier, but it
works.