Add X-Chef-Version HTTP header to Erchef requests
The latest Enterprise Chef release (11.1.0, for those following along
at home) now requires this header for all requests.

The configuration attribute is called "chef_api_version" to both be
descriptive, and to recall the corresponding attribute in Erchef (see
"api_version").  (It's "chef_api_version" to prevent any confusion
that this is describing the Pushy API.)