Return 404 if a principal is requested with a non-existent org
Prior to this commit --
https://github.com/opscode/chef_wm/commit/44517153986dcce444873868929456493b3bb5b3
-- we would return a 403 (Forbidden) if a request was made for a
principal with a non-existent org.  With that chef_wm change in place
(and prior to *this* patch), Pushy would return a 401 (Unauthorized)
for the same request.

This chef_wm change was made in a refactoring pass, and brings more
consistency to the Erchef endpoints.  After much discussion, it was
decided that the 403 / 401 choice was in fact inappropriate, and that
having Pushy return a 404 for these kinds of requests is the more
suitable option.