Change internal handling of principal endpoint responses
A bit of history:

Prior to this chef_wm commit
(https://github.com/opscode/chef_wm/commit/44517153986dcce444873868929456493b3bb5b3)
Pushy would return a 403 (Forbidden) if a request was made for a
principal with a non-existent org.  With that chef_wm change in place
(and prior to *this* patch), Pushy would return a 401 (Unauthorized)
for the same request, instead.

This chef_wm change was made in a refactoring pass, and brings more
consistency to the Erchef endpoints.  After much discussion, it was
decided that the 403 / 401 choice was in fact inappropriate, and that
having Pushy return a 404 for these kinds of requests is the more
suitable option.

Further testing revealed that we needed to distinguish between 404s
due to a missing org vs a missing principal (or a principal not in the
given org).  We modified the principals endpoint to add this
information to the 404 response body, which we now inspect.  A 404 due
to a missing / non-member principal should still result in a 401 (Unauthorized) from
Pushy.