Wait for job to reach terminal state before testing
Timing issues sometimes cause a job to still be in the 'voting' state
when the test retrieves it.  As these tests assume voting will have
already happened (and failed), they expect the status to be
'quorum_failed'.

Since we aren't testing transitions here, but terminal states, we'll
just poll the job until it has reached one of those states, and then
proceed with our tests.