Send an "ack" when the client address changes.
The client will now listen for a message on the command
port, since if encryption is failing there is no error returned.
Therefore, the server must send something when the connection
is established.  When a client first appears, the server
sends an "abort", which serves fine.  But if the client
re-connects, there's no particular message being sent, so
we now send an "ack".