Add attribute to allow client visible server name to be set separately from the VIP
https://github.com/chef/omnibus-pushy/issues/69

A push client needs to be able to reach the backend VIP to connect to the push jobs server. This
causes trouble in some network topologies. This is unlike any other part of chef server; everything
else goes through nginx on the front end nodes, but zeromq can't do that.

The server can advertise to clients a different name to use, but this was defaulted to the backend
VIP, which may not be routable from the clients. Making this independently settable lets the user
choose a name or ip that is routable, perhaps via a firewall rule.

This adds an default attribute ['pushy']['opscode-pushy-server']['server_name_for_client'] to allow
access to the server feature. Setting the value:
opscode_pushy_server['server_name_advertised'] = "example.com"
in /etc/opscode-push-jobs-server/opscode-push-jobs-server.rb will adjust this attribute.