Remove typhoeus from oc-pushy-pedant
typhoeus depends on libcurl at runtime. We would like to remove curl
from the build. It appears that typhoeus is actually unused in the code
so we can drop this dependency.

The EventStream class that uses Typhoeus was dead code. This class has
been removed and the EventStreamOld class has been renamed to
EventStream.

Signed-off-by: Steven Danna <steve@chef.io>